Moves all story related functionality, urls, models, tasks and tests to the mapstories app.
