+++ mapstory/views.py
@@ -20,7 +20,6 @@ from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import Group
 from django.core.urlresolvers import reverse
-from django.db.models import F
 from django.http import HttpResponse, HttpResponseRedirect
 from django.http.request import validate_host
 from django.shortcuts import render_to_response, render, redirect, get_object_or_404
@@ -28,7 +27,6 @@ from django.template import RequestContext
 from django.template.response import TemplateResponse
 from django.utils.http import is_safe_url
 from django.utils.translation import ugettext as _
-from django.views.decorators.clickjacking import xframe_options_exempt
 from django.views.generic import TemplateView
 from django.views.generic.detail import DetailView
 from django.views.generic.list import ListView
@@ -43,11 +41,9 @@ from geonode.layers.views import _PERMISSION_MSG_GENERIC, _PERMISSION_MSG_VIEW, 
 from geonode.geoserver.views import layer_acls, resolve_user
 from geonode.layers.views import _resolve_layer
 from geonode.maps.views import clean_config, map_json, snapshot_config, _PERMISSION_MSG_SAVE, _PERMISSION_MSG_LOGIN
-from geonode.maps.views import add_layers_to_map_config
-from geonode.maps.models import MapLayer, MapSnapshot
 from geonode.security.views import _perms_info_json
 from geonode.layers.tasks import delete_layer
-from geonode.utils import GXPLayer, GXPMap, resolve_object, bbox_to_projection, check_ogc_backend
+from geonode.utils import GXPLayer, GXPMap, bbox_to_projection, check_ogc_backend
 from geonode.utils import build_social_links
 from geonode.utils import default_map_config
 from geonode.utils import DEFAULT_TITLE
@@ -59,21 +55,15 @@ from requests import Request
 
 from mapstory.journal.models import JournalEntry
 from mapstory.initiatives.models import InitiativeMembership
-from mapstory.organizations.models import OrganizationMembership, OrganizationMapStory
-from mapstory.forms import KeywordsForm, MetadataForm, PublishStatusForm, DistributionUrlForm
+from mapstory.organizations.models import OrganizationMembership
+from mapstory.forms import KeywordsForm, MetadataForm, DistributionUrlForm
 from mapstory.importers import GeoServerLayerCreator
 from mapstory.mapstories.models import MapStory, Map
 from mapstory.models import GetPage
 from mapstory.models import Leader
 from mapstory.models import NewsItem
 from mapstory.models import get_sponsors, get_images
-from mapstory.search.utils import update_es_index
 from mapstory.utils import has_exception, parse_wfst_response, print_exception
-from mapstory.favorite.utils import get_favorite_info
-from tasks import delete_mapstory
-from mapstory.thumbnails.tasks import create_mapstory_thumbnail_tx_aware
-from django.views.decorators.http import require_http_methods
-from django.views.decorators.csrf import csrf_exempt
 
 logger = logging.getLogger("geonode.layers.views")
 
@@ -109,108 +99,6 @@ class LeaderListView(ListView):
     context_object_name = 'leaders'
     model = Leader
 
-@xframe_options_exempt
-def map_view(request, mapid, snapshot=None, layer_name=None,
-             template='maps/map_view.html'):
-    """
-    The view that returns the map composer opened to
-    the map with the given map ID.
-    """
-    # This is a MapStory specific change.
-    map_obj = _resolve_story(request, mapid, 'base.view_resourcebase', _PERMISSION_MSG_VIEW)
-    # End MapStory Specific Change
-    if 'access_token' in request.session:
-        access_token = request.session['access_token']
-    else:
-        access_token = None
-
-    if snapshot is None:
-        config = map_obj.viewer_json(request.user, access_token)
-    else:
-        config = snapshot_config(snapshot, map_obj, request.user, access_token)
-
-    if layer_name:
-        config = add_layers_to_map_config(
-            request, map_obj, (layer_name, ), False)
-
-    return render(request, template, context={
-        'config': json.dumps(config),
-        'map': map_obj,
-        'preview': getattr(
-            settings,
-            'GEONODE_CLIENT_LAYER_PREVIEW_LIBRARY',
-            'geoext')
-    })
-
-
-def mapstory_view(request, slug, snapshot=None, template='composer_new/composer.html'):
-    """
-    The view that returns the map viewer opened to
-    the mapstory with the given ID.
-    """
-
-    story_obj = _resolve_map(request, slug, 'base.view_resourcebase', _PERMISSION_MSG_VIEW)
-
-    if 'access_token' in request.session:
-        access_token = request.session['access_token']
-    else:
-        access_token = None
-
-    if snapshot is None:
-        config = story_obj.viewer_json(request.user, access_token)
-    else:
-        config = snapshot_config(snapshot, story_obj, request.user)
-
-    config['about']['detail_url'] = slug
-
-    return render_to_response(template, RequestContext(request, {
-        'config': json.dumps(config),
-        'composerMode': False,
-    }))
-
-
-# TODO this should be moved to a mapstory.util
-def _resolve_story(request, id, permission='base.change_resourcebase',
-                   msg=_PERMISSION_MSG_GENERIC, **kwargs):
-    '''
-    Resolve the Map by the provided typename and check the optional permission.
-    '''
-    if id.isdigit():
-        key = 'pk'
-    else:
-        key = 'slug'
-    return resolve_object(request, MapStory, {key: id}, permission=permission,
-                          permission_msg=msg, **kwargs)
-
-# request a (future) story thumbnail to be created
-# we only allow  PUT/POST since this modifies data
-@require_http_methods(["PUT", "POST"])
-@csrf_exempt
-def story_generate_thumbnail(request, storyid):
-    if not request.user.is_authenticated():
-        return HttpResponse(
-                _PERMISSION_MSG_LOGIN,
-                status=401,
-                content_type="text/plain"
-        )
-
-    story_obj = MapStory.objects.get(id=storyid)
-    if not request.user.has_perm('change_resourcebase', story_obj.get_self_resource()):
-        return HttpResponse(
-                _PERMISSION_MSG_SAVE,
-                status=401,
-                content_type="text/plain"
-        )
-    create_mapstory_thumbnail_tx_aware(story_obj,True)
-    return HttpResponse("create story thumbnail task was scheduled for story id="+str(story_obj.id)+", with uuid="+str(story_obj.uuid))
-
-def composer_new_view(request, slug, template='composer_new/composer.html'):
-    story_obj = _resolve_story(request, slug, 'base.change_resourcebase', _PERMISSION_MSG_SAVE)
-    config = story_obj.viewer_json(request.user)
-    return render_to_response(template, RequestContext(request, {
-        'config': json.dumps(config),
-        'story': story_obj
-    }))
 
 @login_required
 def layer_create(request, template='upload/layer_create.html'):
@@ -924,165 +812,6 @@ def layer_detail(request, layername, template='layers/layer_detail.html'):
         context_dict["maps"] = Map.objects.filter(owner=request.user)
     return TemplateResponse(
         request, template, context=context_dict)
-
-def _resolve_map(request, id, permission='base.change_resourcebase',
-                 msg=_PERMISSION_MSG_GENERIC, **kwargs):
-    '''
-    Resolve the Map by the provided typename and check the optional permission.
-    '''
-    if id.isdigit():
-        key = 'pk'
-    else:
-        key = 'slug'
-    map_obj = resolve_object(request, MapStory, {key: id}, permission=permission,
-                          permission_msg=msg, **kwargs)
-    return map_obj
-
-
-def map_detail(request, slug, snapshot=None, template='maps/map_detail.html'):
-    '''
-    The view that show details of each map
-    '''
-    # MapStory specific change from mapid parameter to slug parameter
-    map_obj = _resolve_map(request, slug, 'base.view_resourcebase', _PERMISSION_MSG_VIEW)
-
-    # Update count for popularity ranking,
-    # but do not includes admins or resource owners
-    if request.user != map_obj.owner and not request.user.is_superuser:
-        Map.objects.filter(
-            id=map_obj.id).update(
-            popular_count=F('popular_count') + 1)
-
-    if 'access_token' in request.session:
-        access_token = request.session['access_token']
-    else:
-        access_token = None
-
-    if snapshot is None:
-        config = map_obj.viewer_json(request.user, access_token)
-    else:
-        config = snapshot_config(snapshot, map_obj, request.user, access_token)
-
-    config = json.dumps(config)
-    layers = MapLayer.objects.filter(map=map_obj.id)
-    links = map_obj.link_set.download()
-
-    group = None
-    if map_obj.group:
-        try:
-            group = GroupProfile.objects.get(slug=map_obj.group.name)
-        except GroupProfile.DoesNotExist:
-            group = None
-
-    ###
-    # MapStory Specific Changes
-    ###
-
-    chapters = map_obj.chapters
-    # Overwrite the layers list so that we can add all layers from all chapters.
-    layers = []
-    for chapter in chapters:
-        layers = layers + list(chapter.local_layers)
-
-    keywords = json.dumps([tag.name for tag in map_obj.keywords.all()])
-
-    if request.method == "POST":
-        keywords_form = KeywordsForm(request.POST, instance=map_obj)
-        published_form = PublishStatusForm(instance=map_obj)
-        if 'keywords' in request.POST:
-            if keywords_form.is_valid():
-                new_keywords = keywords_form.cleaned_data['keywords']
-                map_obj.keywords.add(*new_keywords)
-                map_obj.save()
-            published_form = PublishStatusForm(instance=map_obj)
-        elif 'published_submit_btn' in request.POST:
-            published_form = PublishStatusForm(request.POST, instance=map_obj)
-            if published_form.is_valid():
-                published_form.save()
-                map_obj.is_published = published_form.cleaned_data['is_published']
-                map_obj.save()
-            keywords_form = KeywordsForm(instance=map_obj)
-        elif 'add_keyword' in request.POST:
-            map_obj.keywords.add(request.POST['add_keyword'])
-            map_obj.save()
-        elif 'remove_keyword' in request.POST:
-            map_obj.keywords.remove(request.POST['remove_keyword'])
-            map_obj.save()
-    else:
-        keywords_form = KeywordsForm(instance=map_obj)
-        published_form = PublishStatusForm(instance=map_obj)
-
-    map_thumbnail = map_obj.get_thumbnail_url
-    update_es_index(MapStory, MapStory.objects.get(id=map_obj.id))
-
-    # This will get URL encoded later and is used for the social media share URL
-    share_url = "https://%s/story/%s" % (request.get_host(), map_obj.slug)
-    share_title = "%s by %s." % (map_obj.title, map_obj.owner)
-    share_description = map_obj.abstract
-
-    # Check if user is admin in one of those organizations
-    org_stories = OrganizationMapStory.objects.filter(mapstory=map_obj)
-    org_admin_memberships = []
-    memberships = OrganizationMembership.objects.filter(user_id=request.user.pk)
-    for membership in memberships.all():
-        if membership.is_admin:
-            org_admin_memberships.append(membership)
-
-    if len(org_admin_memberships) < 1:
-        org_admin_memberships = None
-
-    ini_memberships = InitiativeMembership.objects.filter(user_id=request.user.pk)
-    ini_admin_memberships = []
-    for m in ini_memberships.all():
-        if m.is_admin:
-            ini_admin_memberships.append(m)
-
-    if len(ini_admin_memberships) < 1:
-        ini_admin_memberships = None
-
-    ###
-    # End MapStory Specific Changes
-    ###
-
-    context_dict = {
-        'config': config,
-        'resource': map_obj,
-        'group': group,
-        'layers': layers,
-        'perms_list': get_perms(request.user, map_obj.get_self_resource()),
-        'permissions_json': _perms_info_json(map_obj),
-        "documents": get_related_documents(map_obj),
-        'links': links,
-        # MapStory Specific Additions
-        'keywords': keywords,
-        'keywords_form': keywords_form,
-        'published_form': published_form,
-        'thumbnail': map_thumbnail,
-        'share_url': share_url,
-        'share_title': share_title,
-        'share_description': share_description,
-        'organizations': org_admin_memberships,
-        'initiatives': ini_admin_memberships,
-    }
-
-    context_dict["preview"] = getattr(
-        settings,
-        'GEONODE_CLIENT_LAYER_PREVIEW_LIBRARY',
-        'geoext')
-    context_dict["crs"] = getattr(
-        settings,
-        'DEFAULT_MAP_CRS',
-        'EPSG:900913')
-
-    if settings.SOCIAL_ORIGINS:
-        context_dict["social_links"] = build_social_links(request, map_obj)
-
-    # Favorites
-    if request.user.is_authenticated():
-        context_dict["favorite_info"] = get_favorite_info(request.user, map_obj)
-
-    return render(request, template, context=context_dict)
-
 
 @login_required
 def layer_remove(request, layername, template='layers/layer_remove.html'):
